+++ src/query/plan/preprocess.hpp
@@ -261,7 +261,7 @@ struct Matching {
   /// Information on used filter expressions while matching.
   Filters filters;
   /// Maps node symbols to expansions which bind them.
-  std::unordered_map<Symbol, std::set<int>> node_symbol_to_expansions{};
+  std::unordered_map<Symbol, std::set<size_t>> node_symbol_to_expansions{};
   /// Maps named path symbols to a vector of Symbols that define its pattern.
   std::unordered_map<Symbol, std::vector<Symbol>> named_paths{};
   /// All node and edge symbols across all expansions (from all matches).
